backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6

# Database and authentication
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# CORS
starlette>=0.27.0